skele animations

+++ Tile.cpp
@@ -61,7 +61,7 @@ void Tile::initTiles(int& levelPosX, int& levelPosY, const Image& image) {
 }
 
 void Tile::draw() {
-	for (int i = 0; i < tileVector.size(); i++) {
+	for (size_t i = 0; i < tileVector.size(); i++) {
 		if (tileVector[i].sprite.getGlobalBounds().intersects(Game::cullingPoint.getGlobalBounds())) {
 			Game::window.draw(tileVector[i].sprite);
 		}
